Add dependencies on d8 and r8 library arguments
d8 and r8 were missing dependencies on jars passed as library arguments.
This normally worked because the input jar is likely to also depend on
the same jars. When the classpath was added to d8 it started to fail on
libraries that have no code but include other static libraries.

Bug: 116840360
Test: rm out/soong/.intermediates/external/nist-sip/nist-sip/android_common/turbine-combined/nist-sip.jar && m out/soong/.intermediates/frameworks/base/ext/android_common/dex/ext.jar
Change-Id: Ie642f4a343b949778d8ee6dcc4511f3fd7633be1
Signed-off-by: mydongistiny <jaysonedson@gmail.com>